rename 'Active Projects' page to Page 4 and rename 'My Projects Dashboard' to Page 5, since all projects an be found on the Projects page anyway
