+++ src/Pages/Page4.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import GetMyProjects from '../services/notion/getFromNotion/projects/GetActiveProjects';
+import GetMyProjects from '../services/notion/getFromNotion/projects/GetActiveProjects.js';
 import ProgressBar from '../components/ProgressBar/ProgressBar.js';
 
 export default function Page4 () {
